feat(core): self-rotate the S3 OIDC issuer publisher credential
Extend publisher credential rotation to the s3 publisher (after gcs). On the
configured cadence the active node creates a new IAM access key for its own
user, verifies it (STS GetCallerIdentity), swaps it into the live S3 client,
and deletes the old one — so the static IAM key no longer sits unrotated.

- RotatablePublisher on s3Publisher, mirroring the AWS source driver's IAM
  two-key rotation: ListAccessKeys orphan-prune at the 2-key cap, CreateAccessKey,
  verify-with-retry, quota-guard delete-on-verify-failure, in-place client swap
  under a mutex, cleanup/rollback authenticating as the old (propagated) key.
- SupportsRotation gates on the AKIA prefix; config-write rejects rotation_period
  on a non-AKIA s3 key. rotation_period and endpoint owners now include s3, and
  s3 gains an endpoint override (BaseEndpoint + path-style) for S3-compatible
  stores.
- Generalize the persist compare-and-swap: PrepareRotation now returns prevFields
  and persistRotatedPublisherFields does a credential-format-agnostic CAS via
  publisherFieldValue/setPublisherField, so a new publisher type is a data change.
- Bound the IAM list/create sequence with a timeout; tolerate NoSuchEntity on
  deletes; wrap the verify-timeout error for the read API's last_error.
- Split oidc_publisher.go into a shared core plus one file per publisher type,
  with a matching test file per publisher (publish + rotation + related tests).
