+++ cmd/oidcissuer/set_publisher.go
@@ -40,16 +40,21 @@ Usage: warden oidc-issuer set-publisher -type=<type> [-config=key=value ...]
 
       local_file   dir
       http_put     base_url, auth_header, auth_value
-      s3           bucket, region, access_key_id, secret_access_key, prefix
+      s3           bucket, region, access_key_id, secret_access_key, prefix, endpoint
       azure_blob   account_name, container, account_key, prefix, endpoint
       gcs          bucket, credentials_json, prefix, endpoint
       none         (removes the publisher)
 
-  Automatic credential rotation (gcs only) is enabled with -rotation-period: on
-  that cadence Warden mints a fresh service-account key, verifies it, swaps it in,
-  and deletes the old one (the service account must hold
-  iam.serviceAccountKeys.create/delete on itself). Set -rotation-period=0 to
-  disable.
+  Automatic credential rotation (gcs and s3) is enabled with -rotation-period: on
+  that cadence Warden mints a fresh write key, verifies it, swaps it in, and deletes
+  the old one. gcs needs iam.serviceAccountKeys.create/delete on its own service
+  account; s3 needs iam:CreateAccessKey/DeleteAccessKey/ListAccessKeys on its own IAM
+  user (a permanent AKIA… key, not temporary STS credentials). Set -rotation-period=0
+  to disable.
+
+  The s3 endpoint key targets an S3-compatible object store (MinIO, Cloudflare R2,
+  …). Rotation always uses AWS IAM, so -rotation-period applies to real-AWS keys
+  only — leave it unset when publishing to a non-AWS store.
 
   S3:
 
@@ -91,7 +96,7 @@ func init() {
 	f := SetPublisherCmd.Flags()
 	f.StringVar(&pubType, "type", "", "Publisher type: none, local_file, http_put, s3, azure_blob, or gcs")
 	f.StringToStringVar(&pubConfig, "config", nil, "Type-specific config (key=value; repeatable). Use @file to read a value from a file, e.g. -config=account_key=@/run/secrets/key")
-	f.StringVar(&pubRotationPeriod, "rotation-period", "", "Automatic credential-rotation cadence for the publisher's write key (gcs only). A Go duration like 720h; 0 disables. Empty leaves it unchanged.")
+	f.StringVar(&pubRotationPeriod, "rotation-period", "", "Automatic credential-rotation cadence for the publisher's write key (gcs and s3). A Go duration like 720h; 0 disables. Empty leaves it unchanged.")
 	f.StringVarP(&pubJSON, "json", "j", "", "Full publisher JSON block — '<json>', '@file.json', or '-' for stdin (mutually exclusive with -type/-config)")
 }
